% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{binarizeMedian}
\alias{binarizeMedian}
\title{Generate a vector with "High" or "Low" values depending on the original
value compared to the median of the data.}
\usage{
binarizeMedian(data)
}
\arguments{
\item{data}{A numeric dataframe.}
}
\value{
A string dataframe with only 2 possible values.
}
\description{
Generate a vector with "High" or "Low" values depending on the original
value compared to the median of the data.
}
