% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{f_calculateCrossCorrelation}
\alias{f_calculateCrossCorrelation}
\title{Creating cross-correlation profile, phantom peak and calcualting derived QC-metrics}
\usage{
f_calculateCrossCorrelation(data, binding.characteristics, read_length = 36,
  plotname = file.path(getwd(), "phantomCrossCorrelation.pdf"),
  cluster = NULL)
}
\arguments{
\item{data}{Spp-package based object with tag information from bam or tagAlign file(!!!DESCRIBE BETTER)}

\item{binding.characteristics}{(!!!DESCRIBE BETTER)}

\item{read_length}{Integer, read length of "data" (Defaul="36")}

\item{plotname}{Name and path were the CrossCorrelation plot (pdf) should be stored (by DEFAULT stored as "phantomCrossCorrelation.pdf" under the working directory)}

\item{cluster}{Integer, indicating the number of CPUs to parallelize a few functions (default=NULL)}
}
\value{
finalList containing (!!!DESCRIBE BETTER)
strandShift Integer strand shift value from the cross correlation profile
tag.shift Integer rounded halfsize of strandShift
N1 
Nd 
PBC
UNIQUE_TAGS_LibSizeadjusted
phantomScores
STATS_NRF
}
\description{
We use cross-correlation analysis to obtain QC-metrics proposed for narrow-binding patterns. 
After calculating the strand cross-correlation coefficient (Kharchenko et al., 2008), we take the following values from the profile: the coordinates of the 
ChIP-peak (fragment length, height A), the coordinates at the phantom-peak (read length, height B) and the baseline (C), the strand-shift, the number of uniquely mapped 
reads (unique_tags), uniquely mapped reads corrected by the library size, the number of reads and the read lengths. We  calculate different values using the relative and 
absolute height of the cross-correlation peaks: the relative and normalized strand coefficient RSC and NSC  (Landt et al., 2012), and the quality control tag (Fig. 1B) 
(Marinov et al., 2013). Other values regarding the library complexity (Landt et al., 2012) like the fraction of non-redundant mapped reads 
(NRF; ratio between the number of uniquely mapped reads divided by the total number of reads), the NRF adjusted by library size and ignoring the strand direction 
(NRF_nostrand), and the PCR bottleneck coefficient PBC (number of genomic locations to which exactly one unique mapping read maps, divided by the number of unique 
mapping reads). 

f_calculateCrossCorrelation
}
\examples{
\dontrun{
CC_Result=f_CrossCorrelation(chipName, inputName, read_length=36, reads.aligner.type="bam", path=path, dataPath=dataPath, debug=debug,cluster=cluster,chrominfo_file=chrominfo_file)
}
}
