% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_QCscores_local.R
\name{f_CreateMetageneProfile}
\alias{f_CreateMetageneProfile}
\title{Wrapper function to create scaled and non-scaled metageneprofiles}
\usage{
f_CreateMetageneProfile(smoothed.densityChip, smoothed.densityInput, tag.shift,
  annotationID, debug = FALSE)
}
\arguments{
\item{smoothed.densityChip}{DESCRIBE}

\item{smoothed.densityInput}{DESCRIBE}

\item{tag.shift}{Integer, tag shift returned by f_CrossCorrelation()}

\item{annotationID}{String, indicating the genome assembly (Default="hg19")}

\item{debug}{Boolean to enter in debugging mode (default= FALSE)}
}
\value{
list with 3 objects: scaled profile ("twopoint"), non-scaled profile for TSS (TSS) and TES (TES). Each object is made of a list containing 
the chip and the input profile
}
\description{
Metagene plots show the signal enrichment around a region of interest like the TSS or over a predefined set of genes.
The tag density of the immunoprecipitation is taken over all RefSeg annotated human genes, averaged and log2 transformed. The same is done for the 
input. The normalized profile is calculated as the signal enrichment (immunoprecipitation over the input). 
We created two types of metagene profiles: a non-scaled profile for the TSS  and TES, and a scaled profile 
for the entire gene, including the gene body. 
The non-scaled profile is constructed around the TSS/TES, with 2KB up- and downstream regions respectively. 

f_CreateMetageneProfile
}
\examples{
\\{dontrun
Meta_Result=f_CreateMetageneProfile(smoothedDensityChip,smoothedDensityInput,tag.shift,annotationID="hg19",debug=FALSE)
}
}
