% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_QCscores_TFbased.R
\name{f_CrossCorrelation}
\alias{f_CrossCorrelation}
\title{Wrapper function for calculating cross-correlation analysis and metrics designed for TFs}
\usage{
f_CrossCorrelation(chipName, inputName, read_length,
  reads.aligner.type = "bam", dataPath = getwd(), savePlotPath = NULL,
  debug = FALSE, cluster = NULL, mc = 1, chrominfo_file = pwd())
}
\arguments{
\item{chipName}{String, filename (without extension) of the input file for ChIP}

\item{inputName}{String, filename (without extension) of the input file for ChIP}

\item{read_length}{Integer, length of the reads}

\item{reads.aligner.type}{String, indicating the aligner type. Can be "bam" or "tagAlign" (Default="bam")}

\item{dataPath}{Path were to find the input files chipName and inputName, default is working directory}

\item{debug}{Boolean value to enter in debugging mode (default= FALSE)}

\item{cluster}{Integer indicating the number of CPUs to parallelize a few functions (default=NULL)}

\item{chrominfo_file}{Path to the chromatin information file}
}
\value{
returnList, containing (!!!DESCRIBE BETTER)
QCscores_ChIP List with Crosscorrelation values of the ChIP
QCscores_Input List with Crosscorrelation values of the Input
QCscores_binding List with QCscores from called peaks
TagDensityChip (!!!DESCRIBE BETTER)
TagDensityInput (!!!DESCRIBE BETTER)
}
\description{
We use cross-correlation analysis to obtain QC-metrics proposed for narrow-binding patterns. 
After calculating the strand cross-correlation coefficient (Kharchenko et al., 2008), we take the following values from the profile: the coordinates of the 
ChIP-peak (fragment length, height A), the coordinates at the phantom-peak (read length, height B) and the baseline (C), the strand-shift, the number of uniquely mapped 
reads (unique_tags), uniquely mapped reads corrected by the library size, the number of reads and the read lengths. We  calculate different values using the relative and 
absolute height of the cross-correlation peaks: the relative and normalized strand coefficient RSC and NSC  (Landt et al., 2012), and the quality control tag (Fig. 1B) 
(Marinov et al., 2013). Other values regarding the library complexity (Landt et al., 2012) like the fraction of non-redundant mapped reads 
(NRF; ratio between the number of uniquely mapped reads divided by the total number of reads), the NRF adjusted by library size and ignoring the strand direction 
(NRF_nostrand), and the PCR bottleneck coefficient PBC (number of genomic locations to which exactly one unique mapping read maps, divided by the number of unique 
mapping reads). Other measures we include in our analysis are the fraction of usable reads in the peak regions (FRiP) (Landt et al., 2012), for which the function calls sharp- and 
broad-binding peaks to obtain two types: the FRiP_sharpsPeak and the FRiP_broadPeak. The function takes the number of called of peaks using an FDR of 0.01 and an 
evalue of 10 (Kharchenko et al., 2008). And count the number of peaks called when using the sharp- and broad-binding option. 

f_CrossCorrelation
}
\examples{
\\{dontrun
CC_Result=f_CrossCorrelation(chipName, inputName, read_length=36, reads.aligner.type="bam", dataPath=dataPath, debug=debug,cluster=cluster,chrominfo_file=chrominfo_file)
}
}
