% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsGlobal.R
\name{f_fingerPrintPlot}
\alias{f_fingerPrintPlot}
\title{Finger Print Plot}
\usage{
f_fingerPrintPlot(cumChip, cumInput, plotname = file.path(getwd(),
  "FingerPrintPlot.pdf"))
}
\arguments{
\item{cumChip}{The cumulative distribution of the read counts for the ChIP}

\item{cumInput}{The cumulative distribution of the read counts for the Input}

\item{plotname}{Name of the FingerPring plot (default is "chancePlot.pdf" and saved in the working directory )}
}
\value{
returnList
}
\description{
The plot function creates the "Fingerprint plot" i.e. the cumulative distribution of the read counts across genomic bins for Input and ChIP.

f_fingerPrintPlot
}
\examples{
\dontrun{
print("create pdf with plot")
f_fingerPrintPlot(cumSumChip,cumSumInput,plotname=paste(file.path(getwd(),"FingerPrintPlot.pdf",sep="_")
}
}
