% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{f_getBindingRegionsScores}
\alias{f_getBindingRegionsScores}
\title{Calculating QC-values from peak calling procedure}
\usage{
f_getBindingRegionsScores(chip, input, chip.dataSelected, input.dataSelected,
  tag.shift = 75, cluster = NULL)
}
\arguments{
\item{chip}{(!!!DESCRIBE BETTER)}

\item{input}{(!!!DESCRIBE BETTER)}

\item{chip.dataSelected}{(!!!DESCRIBE BETTER)}

\item{input.dataSelected}{(!!!DESCRIBE BETTER)}

\item{cluster}{Integer indicating the number of CPUs to parallelize a few functions (default=NULL)}

\item{tag.shift=75}{(!!!DESCRIBE BETTER)}
}
\value{
QCscoreList, containing (!!!DESCRIBE BETTER)
FDR_detected"=round(FDR_detect,3),
	"eval_detected"=round(eval_detect,3),
	"FRiP_broadPeak"=round(FRiP_broadPeak,3),  
	"FRiP_sharpPeak"=round(FRiP_sharpPeak,3), 
	"outcountsBroadPeak"=outcountsBroadPeak,
	"outcountsSharpPeak"= outcountsSharpPeak)
}
\description{
Other measures we include in our analysis are the fraction of usable reads in the peak regions (FRiP) (Landt et al., 2012), for which the function calls sharp- and 
broad-binding peaks to obtain two types: the FRiP_sharpsPeak and the FRiP_broadPeak. The function takes the number of called of peaks using an FDR of 0.01 and an 
evalue of 10 (Kharchenko et al., 2008). And count the number of peaks called when using the sharp- and broad-binding option. 
f_getBindingRegionsScores
}
\examples{
\\{dontrun
DESCRIBE HERE
}
}
