% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_plot_TSS_TES_allGenes.R
\name{f_plotMetageneProfile_onePoint}
\alias{f_plotMetageneProfile_onePoint}
\title{f_plotMetageneProfile_onePoint}
\usage{
f_plotMetageneProfile_onePoint(binnedChip, binnedInput, tag = "TSS",
  path = getwd(), debug = FALSE, plotName = "NA")
}
\arguments{
\item{chrominfo_file}{}
}
\value{
returnList
}
\description{
f_plotMetageneProfile_onePoint
}
