% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMetaGeneProfilePlots_ForComparison.R
\name{f_plotReferenceDistribution}
\alias{f_plotReferenceDistribution}
\title{Function to create reference distribution plot for comparison}
\usage{
f_plotReferenceDistribution(chrommark, valueToBePlotted = "RSC", currentValue,
  savePlotPath = NULL)
}
\arguments{
\item{valueToBePlotted}{Default="RSC"}

\item{savePlotPath}{Default=NULL, when set saves the density plot as pdf under the gieven path}
}
\description{
Creates a density plot (in pdf) for the sample against the reference distribution (density plots) of the compendium values stratified by chromatin marks.


f_metagenePlotsForComparison
}
\examples{
\\{dontrun
WHATEVER
}
}
