% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_QCscores_global.R
\name{f_QCscores_global}
\alias{f_QCscores_global}
\title{Metrics taken from global read distribution}
\usage{
f_QCscores_global(densityChip, densityInput, savePlotPath = NULL,
  debug = FALSE)
}
\arguments{
\item{densityChip}{A smoothed tag density object for the ChIP (returned by f_CrossCorrelation)}

\item{densityInput}{smoothed tag density object for the Inpt (returned by f_CrossCorrelation)}

\item{debug}{Only for debugging}

\item{plotname}{Name of the FingerPring plot (default is "chancePlot.pdf" and saved in the working directory )}
}
\value{
List with the following values: 
X-axis, Y-Input,Y-Chip (the point of the maximum distance between the ChIP and the input (x-axis, y-axis for immunoprecipitation and input) ,
Ch_sign_chipVSinput (the sign of the maximum distance) 
FractionReadsTopbins_chip, FractionReadsTopbins_input, Fractions_without_reads_chip, Fractions_without_reads_input
}
\description{
This set of values is based on the global read distribution along the genome for immunoprecipitation and input data (Diaz et al., 2012). 
The genome is binned and the read coverage counted for each bin. Then the function computes the cumulative distribution of reads density per genomic bin and plots the 
fraction of the coverage on the y-axis and the fraction of bins on the x-axis. Then different values can be sampled from the cumulative distribution: 
like the fraction of bins without reads for in immunoprecipitation and input,the point of the maximum distance between the ChIP and the input (x-axis, y-axis for 
immunoprecipitation and input, distance (as absolute difference), the sign of the differences), the fraction of reads in the top 1%bin for immunoprecipitation and input. 
Finally, the funciton returns 9 QC-measures

f_QCscores_global
}
\examples{
\dontrun{
CC_Result=f_CrossCorrelation(chipName, inputName, read_length=36, reads.aligner.type="bam", path=path, dataPath=dataPath, debug=debug,cluster=cluster,chrominfo_file=chrominfo_file,plotname=file.path(getwd(),paste(chipName,"CrossCorelationPlot.pdf",sep="_")))
smoothedDensityInput=CC_Result$TagDensityInput
smoothedDensityChip=CC_Result$TagDensityChip
plotname=file.path(getwd(),paste(chipName,"FingerPrintPlot.pdf",sep="_"))
Ch_Results=f_QCscores_global(densityChip=smoothedDensityChip,densityInput=smoothedDensityInput,plotname=plotname,debug=FALSE)
}
}
