% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listFunctions.R
\name{listAvailableElements}
\alias{listAvailableElements}
\title{Shows available chromatin marks and factors}
\usage{
listAvailableElements(target)
}
\arguments{
\item{target}{String, chromatin mark or transcription factor to be analysed.
With the keywords "mark" and "TF" the respective lists with the available 
elements are listed.}
}
\value{
List of elements or single string
}
\description{
Lists chromatin marks and transcription factors that are available 
to be used for the comparison analysis by the fuctions 
metagenePlotsForComparison(), plotReferenceDistribution() 
and predictionScore().

listAvailableElements
}
\examples{

listAvailableElements(target="CTCF")
listAvailableElements(target="H3K36me3")
listAvailableElements(target="TF")
listAvailableElements(target="mark")

}
