% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalFunctions.R
\name{listDatasets}
\alias{listDatasets}
\title{Lists the IDs of samples included in the compendium}
\usage{
listDatasets(dataset)
}
\arguments{
\item{dataset}{String, to specify the dataset for which the IDs 
have to be returned. Valid keywords are "ENCODE" and "Roadmap".}
}
\value{
"ENCODE" returns a vecor of transcription factor, chromatin mark  
and RNAPol2 sample IDs from ENCODE, "Roadmap" returns a vector of 
chromatin mark IDs from Roadmap that have been included in the compendium.
}
\description{
Shows the IDs of all analysed ChIP-seq samples 
included in the compendium from ENCODE and Roadmap. 

listDatasets
}
\examples{

listDatasets(dataset="ENCODE")
listDatasets(dataset="Roadmap")

}
