% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listFunctions.R
\name{listMetrics}
\alias{listMetrics}
\title{Lists the metrics available in the compendium}
\usage{
listMetrics(category = "all")
}
\arguments{
\item{category}{String, to specify the category for which the list of metrics 
should to be returned. Valid keywords are "all", "EM", "GM", "LM".}
}
\value{
returns a character vecor of metrics available in the
compendium that can be used with plotReferenceDistribution() as "metricToBePlotted" parameter
}
\description{
Lists the metrics available in the compendium that can be used with plotReferenceDistribution()

listMetrics
}
\examples{

listMetrics(category="all")
listMetrics(category="EM")

}
