% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMetaGeneProfilePlots_ForComparison.R
\name{plotPredictionScore}
\alias{plotPredictionScore}
\title{Predict score}
\usage{
plotPredictionScore(chrommark = "H3K36me3", features_cc, features_global,
  features_local_TSS, features_local_TES, features_local_scaled,
  savePlotPath = NULL)
}
\arguments{
\item{chrommark}{String, chromatin mark to be analysed. Has to be 
one of the following: H3K36me3, H3K4me3, H3K79me2, H4K20me1,H2AFZ,
H3K27me3, H3K9me3,H3K27ac, POLR2AphosphoS5, H3K9ac, H3K4me2, H3K9me1,
H3K4me1, H3K79me1, H3K4ac, 
H3K14ac, H2BK5ac, H2BK120ac, H2BK15ac, H4K91ac, H4K8ac, H3K18ac, 
H2BK12ac, H3K56ac, H3K23ac, H2AK5ac, H2BK20ac, H4K5ac, H4K12ac, 
H2A.Z, H3K23me2,
H2AK9ac, H3T11ph. For RNAPOL2 different variants are available: 
POLR2A (for RNAPol2), POLR3G and POLR2AphosphoS2}

\item{features_cc}{list, with QC-metrics returned from 
QCscoresCC_PC()}

\item{features_global}{list, list with QC-metrics returned from 
QCscores_global()}

\item{features_local_TSS}{list, list with QC-metrics returned from 
nonScaledMetageneProfile() with option TSS}

\item{features_local_TES}{list,  list with QC-metrics returned from 
nonScaledMetageneProfile() with option TES}

\item{features_local_scaled}{list, list with QC-metrics returned from 
scaledMetageneProfile()}

\item{savePlotPath}{Default=NULL, when set saves the density plot 
(pdf format) under the given path.}
}
\value{
something something
}
\description{
BLABLA 
plotPredictionScore
}
\examples{
\dontrun{
something somethign
}
}
