% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComparisonFunctions.R
\name{plotReferenceDistribution}
\alias{plotReferenceDistribution}
\title{Function to create reference distribution plot for comparison}
\usage{
plotReferenceDistribution(target, metricToBePlotted = "RSC", currentValue,
    savePlotPath = NULL)
}
\arguments{
\item{target}{String, chromatin mark or transcription factor to be analysed.
Use listAvailableElements() function to check availability.}

\item{metricToBePlotted}{The metric to be plotted (Default='RSC')}

\item{currentValue}{The value of the current sample}

\item{savePlotPath}{if set the plot will be saved under 'savePlotPath'. 
Default=NULL and plot will be forwarded to stdout.}
}
\value{
nothing, creates a figure under 'savePlotPath'
}
\description{
Creates a density plot (in pdf) for the sample against the reference 
distribution (density plots) of the compendium values stratified by 
chromatin marks.

plotReferenceDistribution
}
\examples{
print ('Plot distribution of RSC')
\dontrun{
filepath=tempdir()
setwd(filepath)

plotReferenceDistribution(target="H3K4me1", 
    metricToBePlotted="RSC", currentValue=0.49, savePlotPath=filepath)
}
}
