% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{prePost}
\alias{prePost}
\title{Draw a pre-post line for 2 given dataframes.}
\usage{
prePost(x, y, main)
}
\arguments{
\item{data}{x A numeric dataframe of length n}

\item{data}{y A numeric dataframe of length n}
}
\description{
Draw a pre-post line for 2 given dataframes.
}
