% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalFunctions.R
\name{QCscores_global}
\alias{QCscores_global}
\title{Metrics taken from global read distribution}
\usage{
QCscores_global(densityChip, densityInput, savePlotPath = NULL,
  debug = FALSE)
}
\arguments{
\item{densityChip}{Smoothed tag density object for ChIP (returned 
by f_CrossCorrelation).}

\item{densityInput}{Smoothed tag density object for Inpt (returned 
by f_CrossCorrelation)}

\item{savePlotPath}{Path, path set forces the fingerprint plot to be saved under 
the respective directory. Default=NULL, plot not saved but shown on screen}

\item{debug}{Boolean, to enter debugging mode (default= FALSE)}
}
\value{
finalList, list containing 9 QC-values
}
\description{
This set of values is based on the global read distribution along the genome for 
immunoprecipitation and input data (Diaz et al., 2012). The genome is binned and 
the read coverage counted for each bin. Then the function computes the cumulative 
distribution of reads density per genomic bin and plots the fraction of the coverage 
on the y-axis and the fraction of bins on the x-axis. Then different values can be 
sampled from the cumulative distribution: like the fraction of bins without 
reads for in immunoprecipitation and input,the point of the maximum distance 
between the ChIP and the input (x-axis, y-axis for immunoprecipitation and input, 
distance (as absolute difference), the sign of the differences), the fraction of 
reads in the top 1%bin for immunoprecipitation and input. Finally, the funciton 
returns 9 QC-measures

QCscores_global
}
\examples{
\dontrun{
CC_Result=crossCorrelation(chipName=chipName,
inputName=inputName, 
read_length=36, 
dataPath=dataDirectory, 
annotationID="hg19",
savePlotPath=getwd())

tag.shift=CC_Result$QCscores_ChIP$tag.shift
smoothedDensityInput=CC_Result$TagDensityInput
smoothedDensityChip=CC_Result$TagDensityChip

Ch_Results=QCscores_global(densityChip=smoothedDensityChip,
densityInput=smoothedDensityInput)
}
}
