% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalFunctions.R
\name{QCscoresCC_PC}
\alias{QCscoresCC_PC}
\title{Wrapper function to calculate QC-metrics from 
cross-correlation analysis, QC-metrics designed for TFs and 
QC-metrics from peak-calls}
\usage{
QCscoresCC_PC(chipName, inputName, read_length, dataPath = getwd(),
  annotationID = "hg19", mc = 1, savePlotPath = NULL, debug = FALSE)
}
\arguments{
\item{chipName}{String, filename (without extension) of the ChIP file}

\item{inputName}{String, filename (without extension) of the Input file}

\item{read_length}{Integer, length of the reads}

\item{dataPath}{Path, points to the directory were the bam files are 
stored (default is working directory)}

\item{annotationID}{String, indicating the genome assembly (Default="hg19")}

\item{mc}{Integer, the number of CPUs for parallelization (default=1)}

\item{savePlotPath, }{set if Cross-correlation plot should be saved under 
"savePlotPath". Default=NULL and plot will be forwarded to stdout.}

\item{debug}{Boolean, to enter debugging mode (default= FALSE)}
}
\value{
returnList, contains
QCscores_ChIP List of QC-metrics with crosscorrelation values for the ChIP
QCscores_Input List of QC-metrics with crosscorrelation values for the 
Input
QCscores_binding List of QCscores from peak calls
TagDensityChip Tag-density profile, smoothed by the Gaussian kernel 
(for further details see "spp" package)
TagDensityInput Tag density-profile, smoothed by the Gaussian kernel 
(for further details see "spp" package)
}
\description{
Wrapper function that reads input bam files and 
provides QC-metrics from cross-correlation 
analysis, from peak calling and general metrics like read-length or 
NRF. In total 22 features calculated.

QCscoresCC_PC
}
\examples{
print("Example")
\dontrun{
CC_Result=QCscoresCC_PC(chipName=chipName,
inputName=inputName, read_length=36, dataPath=dataDirectory, 
debug=FALSE, mc=1, annotationID="hg19", savePlotPath=getwd())
}
}
