% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_plot_TSS_TES_allGenes.R
\name{qualityScores_LM}
\alias{qualityScores_LM}
\title{Wrapper function that plots non-scaled profiles for 
TSS or TES and to collects the QC-metrics}
\usage{
qualityScores_LM(data, tag, savePlotPath = NULL, debug = FALSE)
}
\arguments{
\item{data}{metagene-list for input and chip sample for
TSS or TES returned by createMetageneProfile()}

\item{tag}{String that can be "TSS" or "TES",indicating if the TSS or 
the TES profile should be calcualted (Default="TSS")}

\item{savePlotPath}{if set the plot will be saved under 
"savePlotPath". Default=NULL and plot will be forwarded to stdout.}

\item{debug}{Boolean to enter in debugging mode (default= FALSE)}
}
\value{
result Dataframe with QC-values for chip, input and 
normalized metagene profile
}
\description{
The non-scaled profile is constructed around the TSS/TES,
with 2KB up- and downstream regions respectively. Different values 
are taken at the TSS/TES and surroundings with +/-2KB, +/-1KB 
and +/-500 sizes. For all the genomic 
positions, we kept the values for the ChIP and the normalized profile,
as the normalization already contains information from the input. 
Additionally, we calculated for all of the intervals between 
the predefined positions the area under the profile, 
the local maxima (x, y coordinates), the variance, 
the standard deviation and the quantiles at 0%, 25%, 
50% and 75%. In total the function returns 43 QC-metrics.

qualityScores_LM
}
\examples{
data(TSSProfile)
TSS_Plot=qualityScores_LM(data=TSSProfile, tag="TSS")

}
