% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalFunctions.R
\name{removeLocalTagAnomalies}
\alias{removeLocalTagAnomalies}
\title{Removes loval anomalies}
\usage{
removeLocalTagAnomalies(chip, input, chip_b.characteristics,
  input_b.characteristics)
}
\arguments{
\item{chip, }{data-structure with tag information for the 
ChIP (see readBamFile())}

\item{input, }{data-structure with tag information for the 
Input (see readBamFile())}

\item{chip_b.characteristics}{binding.characteristics of the ChIP. 
Is a data-structure containing binding information for binding preak
separation distance and cross-correlation profile 
(see spp::get.binding.characteristics)}

\item{input_b.characteristics, }{binding.characteristics of the Input. 
Is a data-structure containing binding information for binding preak
separation distance and cross-correlation profile 
(see spp::get.binding.characteristics)}
}
\value{
result A list containing filtered data structure for ChIP and Input
}
\description{
The removeLocalTagAnomalies function removes tags from regions 
with extremely high tag counts compared to the
neighbourhood.
'
removeLocalTagAnomalies
}
\examples{
## This command is time intensive to run
ChIC.data::chipBam
ChIC.data::inputBam
\dontrun{
chip_binding.characteristics<-spp::get.binding.characteristics(chipBam,
srange=c(0,500), bin=5,accept.all.tags=TRUE)
print ("calculate binding characteristics Input")
input_binding.characteristics<-spp::get.binding.characteristics(inputBam,
srange=c(0,500), bin=5,accept.all.tags=TRUE)
##get chromosome information and order chip and input by it
chrl_final=intersect(names(chipBam$tags),names(inputBam$tags))
chip.data$tags=chipBam$tags[chrl_final]
chip.data$quality=chipBam$quality[chrl_final]
input.data$tags=inputBam$tags[chrl_final]
input.data$quality=inputBam$quality[chrl_final]
##remove sigular positions with extremely high tag counts with 
##respect to the neighbourhood
selectedTags=removeLocalTagAnomalies(chipBam, inputBam, 
chip_binding.characteristics, input_binding.characteristics)
}
}
