% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalFunctions.R
\name{getPeakCallingScores}
\alias{getPeakCallingScores}
\title{Calculating QC-values from peak calling procedure}
\usage{
getPeakCallingScores(chip, input, chip.dataSelected, input.dataSelected,
  tag.shift = 75, chrorder = NULL)
}
\arguments{
\item{chip}{data-structure with tag information for the 
ChIP (see readBamFile())}

\item{input}{data-structure with tag information for the
Input (see readBamFile())}

\item{chip.dataSelected}{selected ChIP tags after running 
removeLocalTagAnomalies() which removes local tag anomalies}

\item{input.dataSelected}{selected Input tags after running 
removeLocalTagAnomalies() which removes local tag anomalies}

\item{tag.shift}{Integer containing the value of the tag shif, 
calculated by getCrossCorrelationScores(). Default=75}

\item{chrorder}{chromosome order (default=NULL)}
}
\value{
QCscoreList List with 6 QC-values
}
\description{
QC-metrics based on the peak calling are the fraction of usable 
reads in the peak regions (FRiP) (Landt et al., 2012), for 
which the function calls sharp- and broad-binding peaks to 
obtain two types: the FRiP_sharpsPeak and 
the FRiP_broadPeak. The function takes the number of called 
of peaks using an 
FDR of 0.01 and an evalue of 10 (Kharchenko et al., 2008). And 
count the number of peaks called when using the 
sharp- and broad-binding option. 

getPeakCallingScores
}
