% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalFunctions.R
\name{qualityScores_GM}
\alias{qualityScores_GM}
\title{Metrics taken from global read distribution}
\usage{
qualityScores_GM(densityChip, densityInput, savePlotPath = NULL,
  debug = FALSE)
}
\arguments{
\item{densityChip}{Smoothed tag-density object for ChIP (returned 
by qualityScores_EM).}

\item{densityInput}{Smoothed tag density object for Input (returned 
by qualityScores_EM)}

\item{savePlotPath}{if set the plot will be saved under 
"savePlotPath". Default=NULL and plot will be forwarded to stdout.}

\item{debug}{Boolean to enter debugging mode (default= FALSE)}
}
\value{
finalList List with 9 QC-values
}
\description{
This set of values is based on the global read distribution along 
the genome for immunoprecipitation and input data (Diaz et al., 2012). 
The genome is binned and the read coverage counted for each bin. Then 
the function computes the cumulative 
distribution of reads density per genomic bin and plots the 
fraction of the coverage on the y-axis and the fraction of bins on 
the x-axis. Then different values can be 
sampled from the cumulative distribution: like the fraction of 
bins without reads for in immunoprecipitation and input,the point of the 
maximum distance between the ChIP and the input 
(x-axis, y-axis for immunoprecipitation and input, 
distance (as absolute difference), the sign of the differences), 
the fraction of reads in the top 1%bin for immunoprecipitation and 
input. Finally, the funciton returns 9 QC-measures

qualityScores_GM
}
\examples{
print("Example Code")
## This command is time intensive to run
##To run the example code the user must provide 2 bam files: 
##one for ChIP and one for the input". Here we used ChIP-seq 
##data from ENCODE. Two example files can be downloaded as follows
## get bam file
setwd(tempdir())

\dontrun{
##chip data
system("wget 
https://www.encodeproject.org/files/ENCFF000BLL/@download/ENCFF000BLL.bam")
chipName="ENCFF000BLL"
chip.data=readBamFile(chipName)

##input data
system("wget 
https://www.encodeproject.org/files/ENCFF000BLL/@download/ENCFF000BKA.bam")
inputName="ENCFF000BKA"
input.data=readBamFile(inputName)

## calculate binding characteristics 
chip_binding.characteristics<-spp::get.binding.characteristics(chip.data, 
srange=c(0,500), bin=5,accept.all.tags=TRUE)

## calculate binding characteristics 
input_binding.characteristics<-spp::get.binding.characteristics(input.data, 
srange=c(0,500), bin=5,accept.all.tags=TRUE)

##get chromosome information and order chip and input by it
chrl_final=intersect(names(chip.data$tags),names(input.data$tags))
chip.data$tags=chip.datatags[chrl_final]
chip.data$quality=chip.data$quality[chrl_final]
input.data$tags=input.data$tags[chrl_final]
input.data$quality=input.data$quality[chrl_final]

##remove sigular positions with extremely high tag counts with 
##respect to the neighbourhood
selectedTags=removeLocalTagAnomalies(chip.data, input.data, 
chip_binding.characteristics, input_binding.characteristics)
input.dataSelected=selectedTags$input.dataSelected
chip.dataSelected=selectedTags$chip.dataSelected

##get smoothed tagdensity 
smoothedChip=tagDensity(chip.dataSelected, 
tag.shift=82)
smoothedInput=tagDensity(input.dataSelected, 
tag.shift=82)

##Finally run function
Ch_Results=qualityScores_GM(densityChip=smoothedChip,
densityInput=smoothedInput)
}
}
