% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_twopoint_allGenes_plot.R
\name{qualityScores_LMgenebody}
\alias{qualityScores_LMgenebody}
\title{Wrapper function to plot the scaled metagene- profile  
and to collect the QC-metrics}
\usage{
qualityScores_LMgenebody(data, savePlotPath = NULL, debug = FALSE)
}
\arguments{
\item{data}{metagene-list for input and chip sample 
of the genebody profile returned by createMetageneProfile()}

\item{savePlotPath}{if set the plot will be saved under 
"savePlotPath". Default=NULL and plot will be forwarded to stdout.}

\item{debug}{Boolean to enter in debugging mode (default= FALSE)}
}
\value{
returnList
}
\description{
The scaled metagene profile that includes the gene body, 
the signal is captured on a real scale from the TSS and an upstream 
region of 2KB. From the TSS, the gene body is constructed with 0.5KB 
in real scale at the gene start (TSS + 0.5KB) and the gene end 
(TES - 0.5KB), whereas the remaining gene body is 
scaled to a virtual length of 2000. Considering the length 
of these regions, the minimum gene length required is 3KB and 
shorter genes are filtered out. From the profile, we take enrichment 
values at different coordinates: at 
-2KB, at the TSS, inner margin (0.5KB), gene body 
(2KB + 2 * inner margin), 
gene body+1KB. We collect in total 42 QC-metrics from the ChIP and 
normalized profile. 

qualityScores_LMgenebody
}
\examples{
print("Example")
data(geneBodyProfile)
geneBody_Plot=qualityScores_LMgenebody(geneBodyProfile)

}
