% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalFunctions.R
\name{readBamFile}
\alias{readBamFile}
\title{Read bam file}
\usage{
readBamFile(filename, path = getwd())
}
\arguments{
\item{filename, }{name of the bam file to be read (without extension)}

\item{path, }{path showing the directory in which the bam file is stored 
(default= workingdirectory)}
}
\value{
list of lists, list corresponds to a chromosome and contains 
a vecotr of coordinates of the 5' ends of the aligned tags
}
\description{
Reading bam file format

readBamFile
}
\examples{
chipName="ENCFF000BLL"
print(chipName)
\dontrun{
chip.data=readBamFile(chipName,path=getwd())
}
}
