% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalFunctions.R
\name{readBamFile}
\alias{readBamFile}
\title{Read bam file}
\usage{
readBamFile(filename)
}
\arguments{
\item{filename, }{name/path of the bam file to be read (without extension)}
}
\value{
result list of lists, every list corresponds to a chromosome 
and contains 
a vecotr of coordinates of the 5' ends of the aligned tags
}
\description{
Reading bam file format

readBamFile
}
\examples{
##To run the example code the user must provide a bam files:
##Here we use a ChIP-seq sample from ENCODE. The file can be downloaded 
##from https://www.encodeproject.org/files/ENCFF000BLL/'
chipName="ENCFF000BLL"
\dontrun{
chip.data=readBamFile(chipName)
}
}
