% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalFunctions.R
\name{readBamFile}
\alias{readBamFile}
\title{Read bam file}
\usage{
readBamFile(filename)
}
\arguments{
\item{filename, }{name/path of the bam file to be read (without extension)}
}
\value{
result list of lists, every list corresponds to a chromosome and 
contains a vector of coordinates of the 5' ends of the aligned tags.
}
\description{
Reading bam file format

readBamFile
}
\examples{

## To run this example code the user MUST provide a bam file: The user can 
## download a ChIP-seq bam file for example from ENCODE:
## https://www.encodeproject.org/files/ENCFF000BFX/
## and save it in the working directory 

bamID="ENCFF000BFX"
\dontrun{
filepath=tempdir()
setwd(filepath)
system("wget 
https://www.encodeproject.org/files/ENCFF000BFX/@download/ENCFF000BFX.bam")

bamName=file.path(filepath,bamID)
chipBam=readBamFile(bamName)
}
}
