% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalFunctions.R
\name{readBamFile}
\alias{readBamFile}
\title{Read bam file}
\usage{
readBamFile(filename)
}
\arguments{
\item{filename, }{name/path of the bam file to be read (without extension)}
}
\value{
result list of lists, every list corresponds to a chromosome 
and contains a vector of coordinates of the 5' ends of the aligned tags
}
\description{
Reading bam file format

readBamFile
}
\examples{
##To run the example code the user must provide 2 bam files: 
##one for ChIP and one for the input". Here we used ChIP-seq 
##data from ENCODE. Two example files can be downloaded as follows
## get bam file
setwd(tempdir())

\dontrun{
##chip data
system("wget 
https://www.encodeproject.org/files/ENCFF000BLL/@download/ENCFF000BLL.bam")
chipName="ENCFF000BLL"
chip.data=readBamFile(chipName)
}
}
