% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalFunctions.R
\name{removeLocalTagAnomalies}
\alias{removeLocalTagAnomalies}
\title{Removes loval anomalies}
\usage{
removeLocalTagAnomalies(chip, input, chip_b.characteristics,
  input_b.characteristics)
}
\arguments{
\item{chip, }{data-structure with tag information for the 
ChIP (see readBamFile())}

\item{input, }{data-structure with tag information for the 
Input (see readBamFile())}

\item{chip_b.characteristics}{binding.characteristics of the ChIP. 
Is a data-structure containing binding information for binding preak
separation distance and cross-correlation profile 
(see spp::get.binding.characteristics)}

\item{input_b.characteristics, }{binding.characteristics of the Input. 
Is a data-structure containing binding information for binding preak
separation distance and cross-correlation profile 
(see spp::get.binding.characteristics)}
}
\value{
result A list containing filtered data structure for ChIP and Input
}
\description{
The removeLocalTagAnomalies function removes tags from regions 
with extremely high tag counts compared to the
neighbourhood.
'
removeLocalTagAnomalies
}
\examples{
print("Usage example")
\dontrun{
chip.data=readBamFile(chipName)
input.data=readBamFile(inputName)
print("calculate binding characteristics ChIP")
### cross_correlation parameters
estimating_fragment_length_range<-c(0,500)
estimating_fragment_length_bin<-5
chip_binding.characteristics<-get.binding.characteristics(chip.data, 
srange=estimating_fragment_length_range, 
bin=estimating_fragment_length_bin,
accept.all.tags=TRUE)
print("calculate cross correlation QC-metrics for the Chip")
crossvalues_Chip<-getCrossCorrelationScores(chip.data,
chip_binding.characteristics,
read_length=read_length,
savePlotPath=savePlotPath,plotname="ChIP")
print("calculate binding characteristics Input")
input_binding.characteristics<-get.binding.characteristics(input.data, 
srange=estimating_fragment_length_range, 
bin=estimating_fragment_length_bin, 
accept.all.tags=TRUE)
print("calculate cross correlation QC-metrics for the Input")
crossvalues_Input=getCrossCorrelationScores
(input.data,input_binding.characteristics,
read_length=read_length, savePlotPath=savePlotPath, plotname="Input")

selectedTags=removeLocalTagAnomalies(chip.data, input.data,
chip_binding.characteristics,
input_binding.characteristics)
}
}
