% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalFunctions.R
\name{removeLocalTagAnomalies}
\alias{removeLocalTagAnomalies}
\title{Removes loval anomalies}
\usage{
removeLocalTagAnomalies(chip, input, chip_b.characteristics,
  input_b.characteristics)
}
\arguments{
\item{chip, }{data structure with tag information from ChIP file}

\item{input, }{data structure with tag information from Input file}

\item{chip_b.characteristics, }{binding-characteristic of the ChiP (previously calculated using get.binding.characteristics)}

\item{input_b.characteristics, }{binding-characteristic of the Input (previously calculated using get.binding.characteristics)}
}
\value{
list, containing filtered data structure for ChIP and Input
}
\description{
The removeLocalTagAnomalies function removes tags from regions with extremely high tag counts compared to the
neighbourhood.
'
removeLocalTagAnomalies
}
\examples{
\\{dontrun
chip.data=readBamFile(chipName,path=dataPath)
input.data=readBamFile(inputName,path=dataPath)
print("calculate binding characteristics ChIP")
## cross_correlation parameters
estimating_fragment_length_range<-c(0,500)
estimating_fragment_length_bin<-5
chip_binding.characteristics<-get.binding.characteristics(chip.data, srange=estimating_fragment_length_range, bin=estimating_fragment_length_bin,accept.all.tags=T)
print("calculate cross correlation QC-metrics for the Chip")
crossvalues_Chip<-calculateCrossCorrelation(chip.data,chip_binding.characteristics,read_length=read_length,savePlotPath=savePlotPath,plotname="ChIP")
print("calculate binding characteristics Input")
input_binding.characteristics<-get.binding.characteristics(input.data, srange=estimating_fragment_length_range, bin=estimating_fragment_length_bin, accept.all.tags=T)
print("calculate cross correlation QC-metrics for the Input")
crossvalues_Input=calculateCrossCorrelation(input.data,input_binding.characteristics,read_length=read_length,savePlotPath=savePlotPath,plotname="Input")

selectedTags=removeLocalTagAnomalies(chip.data,input.data,chip_binding.characteristics,input_binding.characteristics)
}
}
