% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalFunctions.R
\name{tagDensity}
\alias{tagDensity}
\title{Smoothed tag density}
\usage{
tagDensity(data, tag.shift, annotationID = "hg19", mc = 1)
}
\arguments{
\item{data, }{data-structure with tag information (see readBamFile())}

\item{tag.shift, }{Integer containing the value of the tag shif, 
calculated by getCrossCorrelationScores()}

\item{annotationID}{String, indicating the genome assembly 
(Default="hg19")}

\item{mc}{Integer, the number of CPUs for parallelization (default=1)}
}
\value{
smoothed.density A list of lists, each list corresponds to a 
chromosome and contains a 
vector of coordinates of the 5' ends of the aligned tags
}
\description{
Calculates the smoothed tag density using spp::get.smoothed.tag.density

tagDensity
}
\examples{
print("Example")
## This command is time intensive to run
ChIC.data::chipBam
ChIC.data::inputBam
\dontrun{
chip_binding.characteristics<-spp::get.binding.characteristics(chipBam,
srange=c(0,500), bin=5,accept.all.tags=TRUE)
print ("calculate binding characteristics Input")
input_binding.characteristics<-spp::get.binding.characteristics(inputBam,
srange=c(0,500), bin=5,accept.all.tags=TRUE)
##get chromosome information and order chip and input by it
chrl_final=intersect(names(chipBam$tags),names(inputBam$tags))
chip.data$tags=chipBam$tags[chrl_final]
chip.data$quality=chipBam$quality[chrl_final]
input.data$tags=inputBam$tags[chrl_final]
input.data$quality=inputBam$quality[chrl_final]
##remove sigular positions with extremely high tag counts with 
##respect to the neighbourhood
selectedTags=removeLocalTagAnomalies(chipBam, inputBam, 
chip_binding.characteristics, input_binding.characteristics)

chip.dataSelected=selectedTags$chip.dataSelected
smoothed.densityChip=tagDensity(chip.dataSelected, 
tag.shift=82)
}
}
