% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalFunctions.R
\name{tagDensity}
\alias{tagDensity}
\title{Smoothed tag density}
\usage{
tagDensity(data, tag.shift, annotationID = "hg19", mc = 1)
}
\arguments{
\item{data, }{data structure with tag distribution}

\item{tag.shift, }{Integer of the tag shift value returned by 
calculateCrossCorrelation()}

\item{annotationID}{String, indicating the genome assembly (Default="hg19")}

\item{mc}{Integer, the number of CPUs for parallelization (default=1)}
}
\value{
list of lists, list corresponds to a chromosome and contains a vector of 
coordinates of the 5' ends of the aligned tags
}
\description{
calcualtes the smoothed tag density using spp::get.smoothed.tag.density

tagDensity
}
\examples{
\dontrun{
smoothed.densityChip=tagDensity(chip.dataSelected,final.tag.shift,
annotationID="hg19",mc=5)
}
}
